requirements.txt
# recapisure - project dependencies

Flask==2.3.3
Werkzeug==2.3.7

requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3

PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99
sacremoses>=0.0.53

# Apertus LLM用 (4bit量子化)
accelerate>=0.25.0
bitsandbytes>=0.41.0

python-dotenv>=1.0.0

# WebSocket support for real-time progress
flask-socketio>=5.3.0
python-engineio>=4.8.0
python-socketio>=5.10.0

